GroupProjectApi/src/Test/java/s3_gps_ivanti/business/application/impl/GetApplicationBasicInfoUseCaseImplTest.java
package s3_gps_ivanti.business.application.impl;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import s3_gps_ivanti.business.dtoconvertor.ApplicationDTOConverter;
import s3_gps_ivanti.dto.application.ApplicationBasicInfoDTO;
import s3_gps_ivanti.dto.application.GetAllApplicationsResponseDTO;
import s3_gps_ivanti.repository.ApplicationRepository;
import s3_gps_ivanti.repository.entity.Application;
import s3_gps_ivanti.repository.entity.RatingAnalytics;
import s3_gps_ivanti.repository.entity.User;

import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
class GetApplicationBasicInfoUseCaseImplTest {

    @Mock
    private ApplicationRepository applicationRepository;

    @InjectMocks
    private GetApplicationBasicInfoUseCaseImpl getApplicationBasicInfoUseCase;

    /* todo fixme, these where the unit tests for the now broken code
    @Test
    void getAllApplications_TwoResults() {
        Application application1= Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(false)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        Application application2 = Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(false)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        when(applicationRepository.findAll())
                .thenReturn(List.of(application1, application2));


        GetAllApplicationsResponseDTO actualResult = getApplicationBasicInfoUseCase.getAllApplications();

        List<ApplicationBasicInfoDTO> list = ApplicationDTOConverter.convertListToDTO(List.of(application1, application2));
        GetAllApplicationsResponseDTO expectedResult = GetAllApplicationsResponseDTO.builder()
                .applications(list)
                .build();

        assertEquals(expectedResult, actualResult);

        verify(applicationRepository).findAll();
    }
    @Test
    void getAllApplications_OneResults() {
        Application application1= Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(false)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        Application application2 = Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(true)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        when(applicationRepository.findAll())
                .thenReturn(List.of(application1, application2));


        GetAllApplicationsResponseDTO actualResult = getApplicationBasicInfoUseCase.getAllApplications();

        List<ApplicationBasicInfoDTO> list = ApplicationDTOConverter.convertListToDTO(List.of(application1));
        GetAllApplicationsResponseDTO expectedResult = GetAllApplicationsResponseDTO.builder()
                .applications(list)
                .build();

        assertEquals(expectedResult, actualResult);

        verify(applicationRepository).findAll();
    }
    @Test
    void getAllApplications_NoResults() {
        Application application1= Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(true)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        Application application2 = Application.builder()
                .id("id")
                .creator(User.builder()
                        .id("id")
                        .username("creatorName")
                        .email("email")
                        .password("password")
                        .roles(List.of("role"))
                        .permission("permission")
                        .build())
                .name("name")
                .description("description")
                .icon("icon")
                .screenshots(List.of("one","two","three"))
                .status(true)
                .versions(Collections.emptyList())
                .rating(RatingAnalytics.builder()
                        .oneStar(0)
                        .twoStar(0)
                        .threeStar(0)
                        .fourStar(0)
                        .fiveStar(0)
                        .build())
                .build();

        when(applicationRepository.findAll())
                .thenReturn(List.of(application1, application2));


        GetAllApplicationsResponseDTO actualResult = getApplicationBasicInfoUseCase.getAllApplications();

        List<ApplicationBasicInfoDTO> list = ApplicationDTOConverter.convertListToDTO(Collections.emptyList());
        GetAllApplicationsResponseDTO expectedResult = GetAllApplicationsResponseDTO.builder()
                .applications(list)
                .build();

        assertEquals(expectedResult, actualResult);

        verify(applicationRepository).findAll();
    }
    */
}